middleware fixes
